+++ Source/Shaders/SkyAtmosphereFS.glsl
@@ -32,13 +32,14 @@
  
  // Code:  http://sponeil.net/
  // GPU Gems 2 Article:  http://http.developer.nvidia.com/GPUGems2/gpugems2_chapter16.html
- // HSV <-> RGB conversion with minimal branching: http://lolengine.net/blog/2013/07/27/rgb-to-hsv-in-glsl
+ // HSV/HSB <-> RGB conversion with minimal branching: http://lolengine.net/blog/2013/07/27/rgb-to-hsv-in-glsl
 
-uniform vec3 u_hsvShift; // hue, saturation, value
+#ifdef COLOR_CORRECT
+uniform vec3 u_hsbShift; // hue, saturation, value
+#endif
 
 const float g = -0.95;
 const float g2 = g * g;
-const float epsilon = 1.0e-10;
 const vec4 K_RGB2HSB = vec4(0.0, -1.0 / 3.0, 2.0 / 3.0, -1.0);
 const vec4 K_HSB2RGB = vec4(1.0, 2.0 / 3.0, 1.0 / 3.0, 3.0);
 
@@ -48,14 +49,14 @@ varying vec3 v_mieColor;
 varying vec3 v_toCamera;
 varying vec3 v_positionEC;
 
+#ifdef COLOR_CORRECT
 vec3 rgb2hsb(vec3 rgbColor)
 {
     vec4 p = mix(vec4(rgbColor.bg, K_RGB2HSB.wz), vec4(rgbColor.gb, K_RGB2HSB.xy), step(rgbColor.b, rgbColor.g));
     vec4 q = mix(vec4(p.xyw, rgbColor.r), vec4(rgbColor.r, p.yzx), step(p.x, rgbColor.r));
 
     float d = q.x - min(q.w, q.y);
-    float e = 1.0e-10;
-    return vec3(abs(q.z + (q.w - q.y) / (6.0 * d + e)), d / (q.x + e), q.x);
+    return vec3(abs(q.z + (q.w - q.y) / (6.0 * d + czm_epsilon7)), d / (q.x + czm_epsilon7), q.x);
 }
 
 vec3 hsb2rgb(vec3 hsbColor)
@@ -63,6 +64,7 @@ vec3 hsb2rgb(vec3 hsbColor)
     vec3 p = abs(fract(hsbColor.xxx + K_HSB2RGB.xyz) * 6.0 - K_HSB2RGB.www);
     return hsbColor.z * mix(K_HSB2RGB.xxx, clamp(p - K_HSB2RGB.xxx, 0.0, 1.0), hsbColor.y);
 }
+#endif
 
 void main (void)
 {
@@ -78,14 +80,19 @@ void main (void)
     // compute luminance before color correction to avoid strangely gray night skies
     float l = czm_luminance(rgb);
 
-    // convert rgb color to hsv
-    vec3 hsv = rgb2hsb(rgb);
-    // perform hsv shift
-    hsv.x += u_hsvShift.x; // hue
-    hsv.y = clamp(hsv.y + u_hsvShift.y, 0.0, 1.0); // saturation
-    hsv.z = hsv.z > epsilon ? hsv.z + u_hsvShift.z : 0.0; // brightness
-    // convert shifted hsv back to rgb
-    rgb = hsb2rgb(hsv);
+#ifdef COLOR_CORRECT
+    // convert rgb color to hsb
+    vec3 hsb = rgb2hsb(rgb);
+    // perform hsb shift
+    hsb.x += u_hsbShift.x; // hue
+    hsb.y = clamp(hsb.y + u_hsbShift.y, 0.0, 1.0); // saturation
+    hsb.z = hsb.z > czm_epsilon7 ? hsb.z + u_hsbShift.z : 0.0; // brightness
+    // convert shifted hsb back to rgb
+    rgb = hsb2rgb(hsb);
+
+    // check if correction decreased the luminance to 0
+    l = min(l, czm_luminance(rgb));
+#endif
 
     gl_FragColor = vec4(rgb, min(smoothstep(0.0, 0.1, l), 1.0) * smoothstep(0.0, 1.0, czm_morphTime));
 }